add firsts endpoints

+++ backend/internal/interfaces/http/dto/plan_response.go
@@ -2,7 +2,7 @@ package dto
 
 type PlanResponse struct {
 	Id                string `json:"id"`
-	Utgid             int    `json:"utgid"`
+	Utgid             int64  `json:"utgid"`
 	Date              int64  `json:"date"`
 	CaloriesToConsume int    `json:"calories_to_consume"`
 	CaloriesToBurn    int    `json:"calories_to_burn"`
